backend/users/jwt_middleware.py
from urllib.parse import parse_qs
from channels.middleware import BaseMiddleware
from rest_framework_simplejwt.tokens import UntypedToken
from django.contrib.auth.models import AnonymousUser, User
from django.db import close_old_connections

class JWTAuthMiddleware(BaseMiddleware):
    async def __call__(self, scope, receive, send):
        query_string = scope.get('query_string', b'').decode()
        params = parse_qs(query_string)
        token = (params.get('token') or [None])[0]
        scope['user'] = AnonymousUser()
        if token:
            try:
                UntypedToken(token)
                # Simple verification; map to user via token payload 'user_id'
                import jwt
                from django.conf import settings
                payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"], options={"verify_signature": False})
                user_id = payload.get('user_id')
                if user_id:
                    scope['user'] = await self.get_user(user_id)
            except Exception:
                scope['user'] = AnonymousUser()
        close_old_connections()
        return await super().__call__(scope, receive, send)

    @staticmethod
    async def get_user(user_id):
        try:
            return await User.objects.aget(id=user_id)
        except User.DoesNotExist:
            return AnonymousUser()
