backend/requirements.txt
djangorestframework==3.15.2
Django==5.0.6
channels==4.1.0
channels-redis==4.2.0
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
